Set (extend or shorten) vm expiry
Adds a new /setexpiry endpoint where you specify the vm name, expiry in minutes and optionally a vm count and from_reference.
The expired_time is updated for every vm with the given name.
From reference is either now or created and determines whether expired minutes should be from the current time or the time the vm was created. Now is the default